% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_index.r
\name{n_int_a}
\alias{n_int_a}
\title{Additive Intensity Neighbor Index}
\usage{
patch contains the performance of individuals with a neighbor.
open contains the performance of individual without a neighbor.
}
\arguments{
\item{open}{a vector containing non-negatives real values.}

\item{patch}{a vector containing non-negatives real values.}
}
\value{
a vector containing the index values.
}
\description{
Additive Intensity Neighbor Index
}
\details{
This function computes the Additive Intensity Neighbor Index developed by
Dìaz-Sierra et al. (2016). Positive values indicates a positive net effect of
the neighbor (facilitation) whereas negative values indicates a negative one
(competition).
}
\examples{
n_int_a(open = 1, patch = 4)
n_int_a(open = 0.4, patch = 0.1)

}
\seealso{
n_int_c
}
