% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_index.r
\name{n_int_c}
\alias{n_int_c}
\title{Intensity interaction indices}
\usage{
n_int_c(open, patch)
}
\arguments{
\item{open}{a vector containing non-negatives real values.}

\item{patch}{a vector containing non-negatives real values.}
}
\value{
a vector containing the index values.
}
\description{
Intensity interaction indices
}
\details{
Those two functions computes the intensity interaction indices developed by
Dìaz-Sierra et al. (2016). Those indices ensures symmetry, 
n_int_a provided the additive symmetry version of the intensity interaction
index
n_int_c provided the commutative symmetry version of the intensity interaction
}
\examples{



}
