% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.r
\name{pair_find}
\alias{pair_find}
\title{Find the closest open/patch site pairs}
\usage{
pair_find(x)
}
\arguments{
\item{data}{a data.frame}
}
\value{

}
\description{
Find the closest open/patch site pairs
}
\details{
This function uses the Hungarian algorithm to find the closest
open/patch hole pair while minimizing the sum of the distance between the
pairs [please follow this link to have more
details](https://stackoverflow.com/questions/13961493/finding-the-best-matching-pairwise-points-from-2-vectors).
}
\examples{

}
\seealso{
solve_LSAP
}
