% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{holes_data}
\alias{holes_data}
\title{Plant performance summarized by holes}
\format{A data frame with 3524 rows and 12 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{}
}
\usage{
holes_data
}
\description{
A dataset containing all the performance measurements taken in the field,
there are summarized by hole (3 plants by holes) which is my statistical unit.
}
\details{
*Only this dataset should be use in analysis.*
}
\keyword{datasets}
